vertex-ai/audio-transcribe/requirements.txt
vertexai
pyaudio
google-cloud-translate
google-cloud-speech
googletrans
audioport
google-cloud-texttospeech